requirements.txt
boto3>=1.28.0
datasets>=2.16.0
great-expectations>=1.0.0
huggingface_hub>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
pyarrow>=14.0.0
pytest>=7.4.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
